error-handler: extract message generation
Extract the code that translates error objects into messages so we can reuse
it elsewhere.

Credit: @iarna
PR-URL: https://github.com/npm/npm/pull/10465
Reviewed By: @othiym23
