+++ lib/utils/error-handler.js
@@ -2,6 +2,7 @@
 module.exports = errorHandler
 
 var cbCalled = false
+var util = require('util')
 var log = require('npmlog')
 var npm = require('../npm.js')
 var rm = require('rimraf')
@@ -13,6 +14,7 @@ var rollbacks = npm.rollbacks
 var chain = require('slide').chain
 var writeStream = require('fs-write-stream-atomic')
 var nameValidator = require('validate-npm-package-name')
+var errorMessage = require('./error-message.js')
 
 process.on('exit', function (code) {
   log.disableProgress()
@@ -97,6 +99,7 @@ function exit (code, noLog) {
   }
 }
 
+
 function errorHandler (er) {
   log.disableProgress()
   // console.error('errorHandler', er)
@@ -170,319 +173,10 @@ function errorHandler (er) {
   // just a line break
   if (log.levels[log.level] <= log.levels.error) console.error('')
 
-  switch (er.code) {
-    case 'ECONNREFUSED':
-      log.error('', er)
-      log.error(
-        '',
-        [
-          '\nIf you are behind a proxy, please make sure that the',
-          "'proxy' config is set properly.  See: 'npm help config'"
-        ].join('\n')
-      )
-      break
-
-    case 'EACCES':
-    case 'EPERM':
-      log.error('', er)
-      log.error('', ['\nPlease try running this command again as root/Administrator.'
-                ].join('\n'))
-      break
-
-    case 'ELIFECYCLE':
-      log.error('', er.message)
-      log.error(
-        '',
-        [
-          '',
-          'Failed at the ' + er.pkgid + ' ' + er.stage + " script '" + er.script + "'.",
-          'Make sure you have the latest version of node.js and npm installed.',
-          'If you do, this is most likely a problem with the ' + er.pkgname + ' package,',
-          'not with npm itself.',
-          'Tell the author that this fails on your system:',
-          '    ' + er.script,
-          'You can get their info via:',
-          '    npm owner ls ' + er.pkgname,
-          'There is likely additional logging output above.'
-        ].join('\n')
-      )
-      break
-
-    case 'ENOGIT':
-      log.error('', er.message)
-      log.error(
-        '',
-        [
-          '',
-          'Failed using git.',
-          'This is most likely not a problem with npm itself.',
-          'Please check if you have git installed and in your PATH.'
-        ].join('\n')
-      )
-      break
-
-    case 'EJSONPARSE':
-      log.error('', er.message)
-      log.error('', 'File: ' + er.file)
-      log.error(
-        '',
-        [
-          'Failed to parse package.json data.',
-          'package.json must be actual JSON, not just JavaScript.',
-          '',
-          'This is not a bug in npm.',
-          'Tell the package author to fix their package.json file.'
-        ].join('\n'),
-        'JSON.parse'
-      )
-      break
-
-    // TODO(isaacs)
-    // Add a special case here for E401 and E403 explaining auth issues?
-
-    case 'E404':
-      var msg = [er.message]
-      if (er.pkgid && er.pkgid !== '-') {
-        msg.push('', "'" + er.pkgid + "' is not in the npm registry.")
-
-        var valResult = nameValidator(er.pkgid)
-
-        if (valResult.validForNewPackages) {
-          msg.push('You should bug the author to publish it (or use the name yourself!)')
-        } else {
-          msg.push('Your package name is not valid, because', '')
-
-          var errorsArray = (valResult.errors || []).concat(valResult.warnings || [])
-          errorsArray.forEach(function (item, idx) {
-            msg.push(' ' + (idx + 1) + '. ' + item)
-          })
-        }
-
-        if (er.parent) {
-          msg.push("It was specified as a dependency of '" + er.parent + "'")
-        }
-        msg.push(
-          '\nNote that you can also install from a',
-          'tarball, folder, http url, or git url.'
-        )
-      }
-      // There's no need to have 404 in the message as well.
-      msg[0] = msg[0].replace(/^404\s+/, '')
-      log.error('404', msg.join('\n'))
-      break
-
-    case 'EPUBLISHCONFLICT':
-      log.error(
-        'publish fail',
-        [
-          'Cannot publish over existing version.',
-          "Update the 'version' field in package.json and try again.",
-          '',
-          'To automatically increment version numbers, see:',
-          '    npm help version'
-        ].join('\n')
-      )
-      break
-
-    case 'EISGIT':
-      log.error(
-        'git',
-        [
-          er.message,
-          '    ' + er.path,
-          'Refusing to remove it. Update manually,',
-          'or move it out of the way first.'
-        ].join('\n')
-      )
-      break
-
-    case 'ECYCLE':
-      log.error(
-        'cycle',
-        [
-          er.message,
-          'While installing: ' + er.pkgid,
-          'Found a pathological dependency case that npm cannot solve.',
-          'Please report this to the package author.'
-        ].join('\n')
-      )
-      break
-
-    case 'EBADPLATFORM':
-      log.error(
-        'notsup',
-        [
-          er.message,
-          'Not compatible with your operating system or architecture: ' + er.pkgid,
-          'Valid OS:    ' + er.os.join(','),
-          'Valid Arch:  ' + er.cpu.join(','),
-          'Actual OS:   ' + process.platform,
-          'Actual Arch: ' + process.arch
-        ].join('\n')
-      )
-      break
-
-    case 'EEXIST':
-      log.error(
-        [
-          er.message,
-          'File exists: ' + er.path,
-          'Move it away, and try again.'
-        ].join('\n')
-      )
-      break
-
-    case 'ENEEDAUTH':
-      log.error(
-        'need auth',
-        [
-          er.message,
-          'You need to authorize this machine using `npm adduser`'
-        ].join('\n')
-      )
-      break
-
-    case 'EPEERINVALID':
-      var peerErrors = Object.keys(er.peersDepending).map(function (peer) {
-        return 'Peer ' + peer + ' wants ' +
-          er.packageName + '@' + er.peersDepending[peer]
-      })
-      log.error('peerinvalid', [er.message].concat(peerErrors).join('\n'))
-      break
-
-    case 'ECONNRESET':
-    case 'ENOTFOUND':
-    case 'ETIMEDOUT':
-    case 'EAI_FAIL':
-      log.error(
-        'network',
-        [
-          er.message,
-          'This is most likely not a problem with npm itself',
-          'and is related to network connectivity.',
-          'In most cases you are behind a proxy or have bad network settings.',
-          '\nIf you are behind a proxy, please make sure that the',
-          "'proxy' config is set properly.  See: 'npm help config'"
-        ].join('\n')
-      )
-      break
-
-    case 'ENOPACKAGEJSON':
-      log.error(
-        'package.json',
-        [
-          er.message,
-          'This is most likely not a problem with npm itself.',
-          "npm can't find a package.json file in your current directory."
-        ].join('\n')
-      )
-      break
-
-    case 'ETARGET':
-      msg = [
-        er.message,
-        'This is most likely not a problem with npm itself.',
-        'In most cases you or one of your dependencies are requesting',
-        "a package version that doesn't exist."
-      ]
-      if (er.parent) {
-        msg.push("\nIt was specified as a dependency of '" + er.parent + "'\n")
-      }
-      log.error('notarget', msg.join('\n'))
-      break
-
-    case 'ENOTSUP':
-      if (er.required) {
-        log.error(
-          'notsup',
-          [
-            er.message,
-            'Not compatible with your version of node/npm: ' + er.pkgid,
-            'Required: ' + JSON.stringify(er.required),
-            'Actual:   ' + JSON.stringify({
-              npm: npm.version,
-              node: npm.config.get('node-version')
-            })
-          ].join('\n')
-        )
-        break
-      } // else passthrough
-      /*eslint no-fallthrough:0*/
-
-    case 'ENOSPC':
-      log.error(
-        'nospc',
-        [
-          er.message,
-          'This is most likely not a problem with npm itself',
-          'and is related to insufficient space on your system.'
-        ].join('\n')
-      )
-      break
-
-    case 'EROFS':
-      log.error(
-        'rofs',
-        [
-          er.message,
-          'This is most likely not a problem with npm itself',
-          'and is related to the file system being read-only.',
-          '\nOften virtualized file systems, or other file systems',
-          "that don't support symlinks, give this error."
-        ].join('\n')
-      )
-      break
-
-    case 'ENOENT':
-      log.error(
-        'enoent',
-        [
-          er.message,
-          'This is most likely not a problem with npm itself',
-          'and is related to npm not being able to find a file.',
-          er.file ? "\nCheck if the file '" + er.file + "' is present." : ''
-        ].join('\n')
-      )
-      break
-
-    case 'EMISSINGARG':
-    case 'EUNKNOWNTYPE':
-    case 'EINVALIDTYPE':
-    case 'ETOOMANYARGS':
-      log.error(
-        'typeerror',
-        [
-          er.stack,
-          'This is an error with npm itself. Please report this error at:',
-          '    <http://github.com/npm/npm/issues>'
-        ].join('\n')
-      )
-      break
-
-    case 'EISDIR':
-      log.error(
-        'eisdir',
-        [
-          er.message,
-          'This is most likely not a problem with npm itself',
-          'and is related to npm not being able to find a package.json in',
-          'a package you are trying to install.'
-        ].join('\n')
-      )
-      break
-
-    default:
-      log.error('', er.message || er)
-      log.error(
-        '',
-        [
-          '',
-          'If you need help, you may report this error at:',
-          '    <https://github.com/npm/npm/issues>'
-        ].join('\n')
-      )
-      break
-  }
+  var msg = errorMessage(er)
+  msg.summary.concat(msg.detail).forEach(function (errline) {
+    log.error.apply(log, errline)
+  })
 
   exit(typeof er.errno === 'number' ? er.errno : 1)
 }
